optimate/requirements.txt
anthropic>=0.30.0
openai>=1.0.0
groq>=0.4.0
pydantic>=2.5.0
pandas>=2.0.0
numpy>=1.24.0
pulp>=2.7.0
ortools>=9.8.3296
highspy>=1.7.0
scipy>=1.11.0
click>=8.1.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
